verify if file exists before removing it

+++ tests/test_models/test_base_model.py
@@ -22,10 +22,8 @@ class test_basemodel(unittest.TestCase):
         pass
 
     def tearDown(self):
-        try:
+        if os.path.exists('file.json'):
             os.remove('file.json')
-        except NameError:
-            pass
 
     def test_default(self):
         """ """
